requirements.txt
enum34==1.1.6
paho-mqtt==1.3.1
protobuf==3.5.1
robotframework==3.0.2
robotframework-ride==1.5.2.1
six==1.11.0
typing==3.6.2
